// src/api/models/AuthStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CurrentUser } from './CurrentUser';
import {
    CurrentUserFromJSON,
    CurrentUserFromJSONTyped,
    CurrentUserToJSON,
    CurrentUserToJSONTyped,
} from './CurrentUser';

/**
 * 
 * @export
 * @interface AuthStatus
 */
export interface AuthStatus {
    /**
     * 
     * @type {boolean}
     * @memberof AuthStatus
     */
    loggedIn: boolean;
    /**
     * 
     * @type {CurrentUser}
     * @memberof AuthStatus
     */
    user?: CurrentUser | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof AuthStatus
     */
    availablePlayers?: Array<string>;
}

/**
 * Check if a given object implements the AuthStatus interface.
 */
export function instanceOfAuthStatus(value: object): value is AuthStatus {
    if (!('loggedIn' in value) || value['loggedIn'] === undefined) return false;
    return true;
}

export function AuthStatusFromJSON(json: any): AuthStatus {
    return AuthStatusFromJSONTyped(json, false);
}

export function AuthStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthStatus {
    if (json == null) {
        return json;
    }
    return {
        
        'loggedIn': json['logged_in'],
        'user': json['user'] == null ? undefined : CurrentUserFromJSON(json['user']),
        'availablePlayers': json['available_players'] == null ? undefined : json['available_players'],
    };
}

export function AuthStatusToJSON(json: any): AuthStatus {
    return AuthStatusToJSONTyped(json, false);
}

export function AuthStatusToJSONTyped(value?: AuthStatus | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'logged_in': value['loggedIn'],
        'user': CurrentUserToJSON(value['user']),
        'available_players': value['availablePlayers'],
    };
}