// src/api/models/MapVotePage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * radarvan
 * Stats for generals
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MapVoteOption } from './MapVoteOption';
import {
    MapVoteOptionFromJSON,
    MapVoteOptionFromJSONTyped,
    MapVoteOptionToJSON,
    MapVoteOptionToJSONTyped,
} from './MapVoteOption';

/**
 * 
 * @export
 * @interface MapVotePage
 */
export interface MapVotePage {
    /**
     * 
     * @type {number}
     * @memberof MapVotePage
     */
    playerCount: number;
    /**
     * 
     * @type {boolean}
     * @memberof MapVotePage
     */
    loggedIn: boolean;
    /**
     * 
     * @type {number}
     * @memberof MapVotePage
     */
    voteLimit: number;
    /**
     * 
     * @type {number}
     * @memberof MapVotePage
     */
    vetoLimit: number;
    /**
     * 
     * @type {number}
     * @memberof MapVotePage
     */
    votesUsed: number;
    /**
     * 
     * @type {number}
     * @memberof MapVotePage
     */
    vetoesUsed: number;
    /**
     * 
     * @type {Array<MapVoteOption>}
     * @memberof MapVotePage
     */
    maps?: Array<MapVoteOption>;
}

/**
 * Check if a given object implements the MapVotePage interface.
 */
export function instanceOfMapVotePage(value: object): value is MapVotePage {
    if (!('playerCount' in value) || value['playerCount'] === undefined) return false;
    if (!('loggedIn' in value) || value['loggedIn'] === undefined) return false;
    if (!('voteLimit' in value) || value['voteLimit'] === undefined) return false;
    if (!('vetoLimit' in value) || value['vetoLimit'] === undefined) return false;
    if (!('votesUsed' in value) || value['votesUsed'] === undefined) return false;
    if (!('vetoesUsed' in value) || value['vetoesUsed'] === undefined) return false;
    return true;
}

export function MapVotePageFromJSON(json: any): MapVotePage {
    return MapVotePageFromJSONTyped(json, false);
}

export function MapVotePageFromJSONTyped(json: any, ignoreDiscriminator: boolean): MapVotePage {
    if (json == null) {
        return json;
    }
    return {
        
        'playerCount': json['player_count'],
        'loggedIn': json['logged_in'],
        'voteLimit': json['vote_limit'],
        'vetoLimit': json['veto_limit'],
        'votesUsed': json['votes_used'],
        'vetoesUsed': json['vetoes_used'],
        'maps': json['maps'] == null ? undefined : ((json['maps'] as Array<any>).map(MapVoteOptionFromJSON)),
    };
}

export function MapVotePageToJSON(json: any): MapVotePage {
    return MapVotePageToJSONTyped(json, false);
}

export function MapVotePageToJSONTyped(value?: MapVotePage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'player_count': value['playerCount'],
        'logged_in': value['loggedIn'],
        'vote_limit': value['voteLimit'],
        'veto_limit': value['vetoLimit'],
        'votes_used': value['votesUsed'],
        'vetoes_used': value['vetoesUsed'],
        'maps': value['maps'] == null ? undefined : ((value['maps'] as Array<any>).map(MapVoteOptionToJSON)),
    };
}